Pantalla de preferencias, MVC y cambio tamaño textos listo

+++ Android/src/org/inftel/pasos/controlador/IControlador.java
@@ -6,11 +6,11 @@ public interface IControlador {
 	void setNotifVibrador(Boolean b);
 	void setNotifVoz(Boolean b);
 	void setTema(String t);
-	void setTamTexto(int t);
+	void setTamTexto(float t);
 	void aumentarTexto();
 	void disminuirTexto();
 	Boolean getNotifVibracion();
 	Boolean getNotifVoz();
 	String getTema();
-	int getTamTexto();
+	float getTamTexto();
 }
